Removed club model
Clubs become unnecessary and have been removed in a move towards a more
clearer data model pattern
