--- app/controllers/webhooks.server.controller.js
@@ -1,19 +1,0 @@
-'use strict';
-
-/**
- * Module dependencies.
- */
-exports.standings = function(req, res) {
-	//TODO: process standings from kimono
-	res.sendStatus(200);
-};
-
-exports.clubs = function(req, res) {
-	//TODO: process clubs from kimono
-	res.sendStatus(200);
-};
-
-exports.players = function(req, res) {
-	//TODO: process players from kimono
-	res.sendStatus(200);
-};
