+++ app/routes/core.server.routes.js
@@ -3,9 +3,10 @@
 module.exports = function(app) {
 	// Root routing
 	var core = require('../../app/controllers/core'),
-			api = require('../../app/controllers/webhooks');
+		users = require('../../app/controllers/users');
+
 	app.route('/').get(core.index);
-	app.route('/api/league').post(api.standings);
-	app.route('/api/clubs').post(api.clubs);
-	app.route('/api/players').post(api.players);
+
+	app.route('/reset')
+		.get(users.requiresLogin, users.hasAuthorization(['admin']), core.resetDB);
 };
