--- app/tests/club.server.model.test.js
@@ -1,65 +1,0 @@
-'use strict';
-
-/**
- * Module dependencies.
- */
-var should = require('should'),
-	mongoose = require('mongoose'),
-	User = mongoose.model('User'),
-	Club = mongoose.model('Club');
-
-/**
- * Globals
- */
-var user, club;
-
-/**
- * Unit tests
- */
-describe('Club Model Unit Tests:', function() {
-	beforeEach(function(done) {
-
-		user = new User({
-			firstName: 'Full',
-			lastName: 'Name',
-			displayName: 'Full Name',
-			email: 'test@test.com',
-			username: 'username',
-			password: 'password'
-		});
-
-		user.save(function() {
-			club = new Club({
-				name: 'Club Name',
-				user: user
-			});
-
-			done();
-		});
-	});
-
-	describe('Method Save', function() {
-		it('should be able to save without problems', function(done) {
-			club.name = 'Club';
-			club.whoscoredId = 111;
-			return club.save(function(err) {
-				should.not.exist(err);
-				done();
-			});
-		});
-
-		it('should be able to show an error when try to save without name or whoscoredId', function(done) {
-			return club.save(function(err) {
-				should.exist(err);
-				done();
-			});
-		});
-	});
-
-	afterEach(function(done) { 
-		Club.remove().exec();
-		User.remove().exec();
-
-		done();
-	});
-});
