--- public/modules/clubs/clubs.client.module.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// Use applicaion configuration module to register a new module
-ApplicationConfiguration.registerModule('clubs');
