--- public/modules/clubs/views/list-clubs.client.view.html
@@ -1,19 +1,0 @@
-<section data-ng-controller="ClubsController" data-ng-init="find()">
-    <div class="page-header">
-        <h1>Clubs</h1>
-    </div>
-    <div class="list-group">
-        <a data-ng-repeat="club in clubs" data-ng-href="#!/clubs/{{club._id}}" class="list-group-item">
-			<small class="list-group-item-text">
-				Posted on
-				<span data-ng-bind="club.created | date:'medium'"></span>
-				by
-				<span data-ng-bind="club.user.displayName"></span>
-			</small>
-			<h4 class="list-group-item-heading" data-ng-bind="club.name"></h4>
-        </a>
-    </div>
-    <div class="alert alert-warning text-center" data-ng-hide="!clubs.$resolved || clubs.length">
-    	No Clubs yet, why don't you <a href="/#!/clubs/create">create one</a>?
-    </div>
-</section>
