--- public/modules/clubs/views/view-club.client.view.html
@@ -1,21 +1,0 @@
-<section data-ng-controller="ClubsController" data-ng-init="findOne()">
-	<div class="page-header">
-		<h1 data-ng-bind="club.name"></h1>
-	</div>
-	<div class="pull-right" data-ng-show="authentication.user._id == club.user._id">
-		<a class="btn btn-primary" href="/#!/clubs/{{club._id}}/edit">
-			<i class="glyphicon glyphicon-edit"></i>
-		</a>
-		<a class="btn btn-primary" data-ng-click="remove();">
-			<i class="glyphicon glyphicon-trash"></i>
-		</a>
-	</div>
-	<small>
-		<em class="text-muted">
-			Posted on
-			<span data-ng-bind="club.created | date:'mediumDate'"></span>
-			by
-			<span data-ng-bind="club.user.displayName"></span>
-		</em>
-	</small>
-</section>
